requirements.txt
Flask==3.1.3
zeroconf==0.147.0           # optional: mDNS-Scan für Shelly-Geräte
websocket-client==1.9.0
PyYAML==6.0.3
requests==2.32.5
certifi==2026.2.25          # sauberes CA-Bundle (s. Patch unten)
waitress==3.0.2                             # Prod-WSGI (Windows + Linux/Docker)
# gunicorn==22.0.0 ; sys_platform != "win32" # Alternative für Linux/macOS

# Dev / Test
pytest==8.3.5
pytest-cov==6.1.1